Added pre-compile for HTML page
The HTML page that displays the manga list is not expected
to change that often so it is only necessary to compile it
when new data is fetched.

It is for this reason that we will now compile the html beforehand
and then serve the static file, that way it will be a lot faster.